after migration

+++ lib/models/details_of_shop_owner_model.dart
@@ -1,7 +1,7 @@
 class ProfileShopOwnerSummaryModel {
   String? profileName;
   String? profilePicture;
-  double? overallStoreRating;
+  int? overallStoreRating;
   int? totalStores;
   int? totalProducts;
   int? totalViews;
